Split applyAxis functions into separate namespace
